finish auth

+++ frontend/src/components/CycleDetails.js
@@ -1,13 +1,20 @@
-import React from 'react'
 import { useCyclesContext } from '../hooks/useCyclesContext'
 import formatDistanceToNow from 'date-fns/formatDistanceToNow'
+import { useAuthContext } from '../hooks/useAuthContext';
 
 const CycleDetails = ({cycle}) => {
   const {dispatch} = useCyclesContext();
+  const {user} = useAuthContext();
 
   const handleClick = async () => {
+    if (!user) {
+      return;
+    }
     const response = await fetch('api/cycles/' + cycle._id, {
-      method: 'DELETE'
+      method: 'DELETE',
+      headers: {
+        'Authorization': `Bearer ${user.token}`
+      }
     })
     const json = await response.json()
